pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
